Sum of Double

+++ Calculator.java
@@ -1,9 +1,10 @@
 public class Calculator {
 
-public void sum(int a, int b){
+public void sum(double a, double b){
 	int sum = a + b;
-
+    
 	System.out.println(sum);
+
 }
